Wheels seem to build correctly now
Former-commit-id: f63bda7afdef0e92216a084d3f537efc3df016be [formerly 073cacca47a370c714c580973ce727bb1d90e5c4]
Former-commit-id: ea225f7632f20bb6301c9ee09385b14ff07a1574

+++ src/librapid/array/functors/binary/binary.hpp
@@ -9,7 +9,7 @@ namespace librapid::functors::binary {
 		using RetType				   = std::false_type;
 		static constexpr int64_t Flags = 0;
 
-		template<typename T, int64_t d, int64_t a>
+		template<typename T, i32 d, i32 a>
 		LR_NODISCARD("")
 		ExtentType<T, d> genExtent(const ExtentType<T, d, a> &lhs,
 								   const ExtentType<T, d, a> &rhs) const {
@@ -17,7 +17,7 @@ namespace librapid::functors::binary {
 		}
 
 		// For a scalar operation
-		template<typename T, int64_t d, int64_t a>
+		template<typename T, i32 d, i32 a>
 		LR_NODISCARD("")
 		ExtentType<T, d> genExtent(const ExtentType<T, d, a> &generic) const {
 			return generic;
